requirements.txt
fastapi==0.120.2
uvicorn==0.38.0
requests==2.32.5
python-dotenv==1.2.1
pydantic==2.12.3
starlette==0.49.1